feat: remove z and cosmo as parameters from MassDefinition
The reason for this change is that it's nice to have
very lightweight constructors for MassDefinition in other
places, in order to compare between definitions. Arguably,
two definitions should be considered the same even if they
are at different redshift/cosmology.

Also, it removes the ability to do mass conversion without
halomod installed, so that there's only one place we're defining
the halo profile and concentration.

It also fixes redshift-dependence of mean density in a few places.